Mua hang, ban hang, Kho

+++ Accounting.Domain/Entities/DonMuaDong.cs
@@ -2,9 +2,9 @@
 {
     public class DonMuaDong
     {
-        public int Id { get; set; }
-        public int? DonMuaId { get; set; }
-        public int? VatTuId { get; set; }
+        public long Id { get; set; }
+        public long? DonMuaId { get; set; }
+        public long? VatTuId { get; set; }
         public string? KichThuoc { get; set; }
         public string? LoaiGiay { get; set; }
         public string? DinhLuongGsm { get; set; }
@@ -14,7 +14,7 @@
         public decimal SoLuong { get; set; }
         public decimal DonGia { get; set; }
 
-        public int? ThueSuatId { get; set; }
+        public long? ThueSuatId { get; set; }
         public decimal? TienThue { get; set; }
         public decimal? ThanhTien { get; set; }
     }
