+++ Accounting.Domain/Entities/HoaDonMua.cs
@@ -2,10 +2,10 @@
 {
     public class HoaDonMua
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string SoCt { get; set; } = null!;
-        public int? NhaCungCapId { get; set; }
-        public int? DonMuaId { get; set; }
+        public long? NhaCungCapId { get; set; }
+        public long? DonMuaId { get; set; }
 
         public DateTime NgayHoaDon { get; set; }
         public DateTime? HanThanhToan { get; set; }
