+++ Accounting.Domain/Entities/HoaDonMuaDong.cs
@@ -2,14 +2,14 @@
 {
     public class HoaDonMuaDong
     {
-        public int Id { get; set; }
-        public int? HoaDonMuaId { get; set; }
-        public int? VatTuId { get; set; }
+        public long Id { get; set; }
+        public long? HoaDonMuaId { get; set; }
+        public long? VatTuId { get; set; }
 
         public decimal SoLuong { get; set; }
         public decimal DonGia { get; set; }
 
-        public int? ThueSuatId { get; set; }
+        public long? ThueSuatId { get; set; }
         public decimal? TienThue { get; set; }
         public decimal? ThanhTien { get; set; }
     }
